dbeaver/pro#6770 fix delete

+++ plugins/org.jkiss.dbeaver.ui.dashboard/src/org/jkiss/dbeaver/ui/dashboard/control/DashboardListControl.java
@@ -346,7 +346,9 @@ public class DashboardListControl extends Composite implements DashboardGroupCon
                 itemToDelete = dashContainerItem;
             }
         }
-        deleteItem(itemToDelete);
+        if (itemToDelete != null) {
+            deleteItem(itemToDelete);
+        }
     }
 
     @Override
